--- init.php
@@ -1,5 +1,0 @@
-<?php
-
-require(dirname(__FILE__) . '/lib/Berbix.php');
-
-?>
